+++ NEWS.md
@@ -1,5 +1,6 @@
 # queryparser (development version)
 
+* Column names that match SQL function names are not converted to lowercase ([#1](https://github.com/ianmcook/queryparser/issues/1))
 * Operators `&&` and `||` are not used in outputted expressions ([#2](https://github.com/ianmcook/queryparser/issues/2))
 * Second through last occurrences of binary symbolic operators are replaced ([#3](https://github.com/ianmcook/queryparser/issues/3))
 * Column names that match function names are not replaced ([#4](https://github.com/ianmcook/queryparser/issues/4))